database/requirements.txt
Django==4.0.3
djangorestframework==3.13.1
et-xmlfile==1.1.0
openpyxl==3.0.9
pandas==1.4.1
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
status==0.2.5
urllib3==1.26.8
XlsxWriter==3.0.3
